Resolve "Clean up warnings during compilation of icet_cpp via setup.py"

Closes #293

See merge request materials-modeling/icet!238

+++ src/ClusterExpansionCalculator.hpp
@@ -25,9 +25,10 @@ class ClusterExpansionCalculator
     ClusterExpansionCalculator(const ClusterSpace &, const Structure &);
 
     /// Returns the local cluster vector.
-    std::vector<double> getLocalClusterVector(const std::vector<int>& ,const int, const std::vector<int>);
+    std::vector<double> getLocalClusterVector(const std::vector<int>&, const int, const std::vector<size_t>);
 
     private:
+  
     /// Maps offsets to local orbit lists.
     std::unordered_map<Vector3d, OrbitList, Vector3dHash> _localOrbitlists;
